(#13) refactor: 등록, 수정, 삭제 기능 timeform에서 처리
