+++ src/app/timecheck/page.js
@@ -1,7 +1,6 @@
 import React from "react";
 import s from "./timeCheckPage.module.css";
 import Calender from "@/components/time/calender";
-import Timeform from "@/components/time/timeform";
 
 export const metadata = {
   title: "Read Time | TimeCheck",
@@ -12,7 +11,6 @@ export default function TimeCheck() {
   return (
     <div className={s.contain}>
       <Calender />
-      <Timeform />
     </div>
   );
 }
