+++ src/app/timecheck/timeCheckPage.module.css
@@ -4,7 +4,7 @@
   gap: 2rem;
 }
 
-.contain section {
+.contain>section {
   width: 50%;
 }
