+++ src/components/time/calender.js
@@ -2,34 +2,28 @@
 import React, { useEffect, useState } from "react";
 import Calendar from "react-calendar";
 import moment from "moment/moment";
-import { useSession } from "next-auth/react";
-import TimeEditModal from "./timeeditmodal";
-import s from "./calender.module.css";
+import Timeform from "./timeform";
 
 export default function Calender() {
   const [value, onChange] = useState(new Date());
   const [record, setRecord] = useState();
-  const [clickDay, setClickDay] = useState(false);
-  const { data: session } = useSession();
 
   useEffect(() => {
-    if (session) {
-      const dataFetch = async () => {
-        try {
-          const response = await fetch("/api/getrecord", {
-            cache: "no-store",
-            credentials: "include",
-          });
-          const data = await response.json();
-          setRecord(data);
-          return data;
-        } catch (error) {
-          console.error(error);
-        }
-      };
-      dataFetch();
-    }
-  }, [session]);
+    const dataFetch = async () => {
+      try {
+        const response = await fetch("/api/getrecord", {
+          cache: "no-store",
+          credentials: "include",
+        });
+        const data = await response.json();
+        setRecord(data);
+        return data;
+      } catch (error) {
+        console.error(error);
+      }
+    };
+    dataFetch();
+  }, []);
 
   let highlightList;
   if (record) {
@@ -66,38 +60,37 @@ export default function Calender() {
   };
 
   return (
-    <section>
-      <p className={s.info}>
-        시간 수정과 삭제는 해당 날짜를 눌러 진행해주세요.
-      </p>
-      <Calendar
-        locale="ko"
-        onChange={onChange}
-        value={value}
-        next2Label={null}
-        prev2Label={null}
-        tileContent={addContent}
-        showNeighboringMonth={false}
-        onClickDay={() => setClickDay(true)}
-        formatDay={(locale, date) => moment(date).format("DD")}
-        tileClassName={({ date, view }) => {
-          if (
-            record &&
-            highlightList.find((x) => x === moment(date).format("YYYY-MM-DD"))
-          ) {
-            return "highlight";
-          }
-        }}
-      />
-      {clickDay && (
-        <TimeEditModal
-          setClickDay={setClickDay}
+    <>
+      <section>
+        <Calendar
+          locale="ko"
+          onChange={onChange}
           value={value}
-          record={record.filter(
-            (item) => item.date === moment(value).format("YYYY-MM-DD")
-          )}
+          next2Label={null}
+          prev2Label={null}
+          tileContent={addContent}
+          showNeighboringMonth={false}
+          formatDay={(locale, date) => moment(date).format("DD")}
+          tileClassName={({ date, view }) => {
+            if (
+              record &&
+              highlightList.find((x) => x === moment(date).format("YYYY-MM-DD"))
+            ) {
+              return "highlight";
+            }
+          }}
         />
-      )}
-    </section>
+      </section>
+      <Timeform
+        value={value}
+        record={
+          record
+            ? record.filter(
+                (item) => item.date === moment(value).format("YYYY-MM-DD")
+              )
+            : []
+        }
+      />
+    </>
   );
 }
