--- src/components/time/calender.module.css
@@ -1,5 +1,0 @@
-.info {
-  font-size: 14px;
-  color: var(--gray300-color);
-  margin-bottom: 1rem;
-}
