--- src/components/time/timeeditmodal.module.css
@@ -1,38 +1,0 @@
-.tiemInputDiv {
-  display: flex;
-  justify-content: space-evenly;
-  margin-top: 1rem;
-}
-
-.timeInput input {
-  width: 3rem;
-  padding: 10px 20px;
-  border-bottom: 1px solid var(--gray200-color);
-  display: inline-block;
-  margin-right: 8px;
-}
-
-.btnDiv {
-  margin-top: 1rem;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 8px;
-}
-
-.btnDiv button {
-  padding: 1rem;
-  cursor: pointer;
-  background-color: #1c1c1c;
-  border-radius: 8px;
-  transition: all 0.3s;
-  width: 50%;
-}
-
-.btnDiv button:last-of-type:hover {
-  filter: brightness(150%);
-}
-
-.btnDiv button:first-of-type:hover {
-  color: red;
-}
