+++ src/components/time/timeform.js
@@ -1,10 +1,10 @@
 "use client";
-import React, { useRef, useState } from "react";
+import React, { useRef } from "react";
 import s from "./timeform.module.css";
 import { useSession } from "next-auth/react";
+import moment from "moment";
 
-export default function Timeform() {
-  const [date, setDate] = useState(getCurrentDate());
+export default function Timeform({ value, record }) {
   const hoursRef = useRef();
   const minutesRef = useRef();
   const { data: session } = useSession();
@@ -15,12 +15,14 @@ export default function Timeform() {
     const minutesValue = minutesRef.current.value;
     const formattedHours = hoursValue.toString().padStart(2, "0");
     const formattedMinutes = minutesValue.toString().padStart(2, "0");
+    const formattedReadingTime = `${formattedHours}:${formattedMinutes}`;
 
     if (formattedHours === "00" && formattedMinutes === "00") {
       alert("시간을 입력해주세요!");
-    } else {
-      const formattedReadingTime = `${formattedHours}:${formattedMinutes}`;
+      return;
+    }
 
+    if (!record || record.length === 0) {
       try {
         const response = await fetch("/api/timerecord", {
           method: "POST",
@@ -29,39 +31,68 @@ export default function Timeform() {
           },
           body: JSON.stringify({
             username: session.user.name,
-            date: date,
+            date: moment(value).format("YYYY-MM-DD"),
             time: formattedReadingTime,
           }),
         });
 
         if (response.ok) {
           alert("기록 완료!");
-        } else {
-          alert("기록에 실패했습니다.");
         }
       } catch (error) {
         console.error("데이터 저장 오류:", error);
         alert("저장 중 오류가 발생했습니다.");
-      } finally {
-        hoursRef.current.value = "";
-        minutesRef.current.value = "";
+      }
+    } else {
+      try {
+        const response = await fetch("/api/timerecord/edit", {
+          method: "PATCH",
+          headers: {
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify({
+            _id: record[0]._id,
+            time: formattedReadingTime,
+          }),
+        });
+
+        if (response.ok) {
+          alert("수정 완료!");
+        }
+      } catch (e) {
+        console.error("데이터 저장 오류:", e);
+        alert("저장 중 오류가 발생했습니다.");
       }
     }
+    hoursRef.current.value = "";
+    minutesRef.current.value = "";
   };
 
-  function getCurrentDate() {
-    const today = new Date();
-    const year = today.getFullYear();
-    const month = String(today.getMonth() + 1).padStart(2, "0");
-    const day = String(today.getDate()).padStart(2, "0");
-    return `${year}-${month}-${day}`;
-  }
+  const handleTimeDelete = async () => {
+    try {
+      const response = await fetch("/api/timerecord/remove", {
+        method: "DELETE",
+        body: JSON.stringify({
+          _id: record[0]._id,
+        }),
+      });
+
+      if (response.ok) {
+        alert("삭제 완료!");
+      } else {
+        alert("삭제에 실패했습니다.");
+      }
+    } catch (e) {
+      console.error(e);
+    }
+  };
 
   return (
     <section className={s.timeSection}>
       <h2>오늘의 독서 시간 기록</h2>
       <form onSubmit={handleSubmit} className={s.timeForm}>
-        <p>{date} 기록</p>
+        <p>{moment(value).format("YYYY-MM-DD")} 기록</p>
+        <p className={s.info}>캘린더를 눌러 수정과 삭제를 할 수 있어요.</p>
         <div className={s.tiemInputDiv}>
           <div className={s.timeInput}>
             <input
@@ -86,9 +117,18 @@ export default function Timeform() {
             <label htmlFor="minutes">분</label>
           </div>
         </div>
-        <button type="submit" className={s.timeBtn}>
-          저장하기
-        </button>
+        <div className={s.timeBtnDiv}>
+          {record.length >= 1 && (
+            <button
+              type="button"
+              onClick={handleTimeDelete}
+              className={s.timeRemove}
+            >
+              삭제하기
+            </button>
+          )}
+          <button type="submit">저장하기</button>
+        </div>
       </form>
     </section>
   );
