+++ src/components/time/timeform.module.css
@@ -5,6 +5,12 @@
   border-bottom: 1px solid #777;
 }
 
+.info {
+  font-size: 14px;
+  color: var(--gray300-color);
+  margin-top: 10px;
+}
+
 .timeForm {
   margin-top: 1.5rem;
 }
@@ -13,10 +19,10 @@
   display: flex;
   gap: 2rem;
   align-items: center;
+  margin-top: 20px;
 }
 
 .timeInput {
-  margin-top: 1rem;
   width: 50%;
   display: flex;
   gap: 8px;
@@ -34,12 +40,22 @@
   flex-grow: 0;
 }
 
-.timeBtn {
-  cursor: pointer;
+.timeBtnDiv {
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+  gap: 1rem;
   margin-top: 1rem;
-  display: block;
-  margin-left: auto;
-  padding: 10px 0 10px 2rem;
+}
+
+.timeBtnDiv button {
+  cursor: pointer;
+  padding: 10px 0 10px 1rem;
+  transition: all 0.3s;
+}
+
+.timeRemove:hover {
+  color: red;
 }
 
 @media screen and (max-width: 480px) {
